backend/marketplace/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .api.views import OrderRequestViewSet

router = DefaultRouter()
router.register(r'orders', OrderRequestViewSet, basename='order')

urlpatterns = [
    path('', include(router.urls)),
]