backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
whitenoise>=6.6.0
django-jazzmin>=3.0.0